requirements.txt
langchain>=0.3.0
langchain-mistralai>=0.2.0
langchain-community>=0.3.0
arxiv>=2.1.0
fastapi>=0.115.0
uvicorn>=0.32.0
streamlit>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
